Added dark out boxes
