+++ frontend/src/Student/components/PortfolioCard.js
@@ -8,10 +8,10 @@ import FaPlusCircle from '@fortawesome/fontawesome-free-solid/faPlusCircle'
 import FaBook from '@fortawesome/fontawesome-free-solid/faBook'
 import FaYouTube from '@fortawesome/fontawesome-free-brands/faYoutube'
 import FaVimeo from '@fortawesome/fontawesome-free-brands/faVimeoV'
-import Flippy, { FrontSide, BackSide } from 'react-flippy';
 import ReactCardFlip from 'react-card-flip';
 import { Row, Col, Button } from 'reactstrap'
 import moment from 'moment'
+import {FlipCard, DarkCard} from './ImageCard'
 
 import { getImageThumbnail, STATIC_PATH } from '../../utils'
 
@@ -72,46 +72,6 @@ const NotInvited = styled.div`
   width: inherit;
 `
 
-const EntryThumb = ({ entry }) => {
-  switch (entry.entryType) {
-    case 'PHOTO':
-      return (
-        <PhotoThumbnail
-          src={`${STATIC_PATH}${getImageThumbnail(entry.path)}`}
-        />
-      )
-    case 'VIDEO':
-      const url =
-        entry.provider === 'youtube'
-          ? `https://youtube.com/watch?v=${entry.videoId}`
-          : `https://vimeo.com/${entry.videoId}`
-      const icon =
-        entry.provider === 'youtube' ? (
-          <FontAwesomeIcon icon={FaYouTube} size='3x' />
-        ) : (
-          <FontAwesomeIcon icon={FaVimeo} size='3x' />
-        )
-      return (
-        <a href={url} target='_blank'>
-          <EntryNoThumbContainer>
-            {icon}
-            <h5>{entry.title}</h5>
-          </EntryNoThumbContainer>
-        </a>
-      )
-    case 'OTHER':
-      return (
-        <EntryNoThumbContainer>
-          <FontAwesomeIcon icon={FaBook} size='3x' />
-          <h5>{entry.title}</h5>
-        </EntryNoThumbContainer>
-      )
-    default:
-      return null
-  }
-}
-
-
 const NewSubmission = ({ show }) => (
   <Col
     style={{ minHeight: '10em' }}
@@ -125,45 +85,6 @@ const NewSubmission = ({ show }) => (
   </Col>
 )
 
-class FlipCard extends React.Component {
-  constructor() {
-    super();
-      this.state = {
-      isFlipped: false
-    };
-    this.flip = this.flip.bind(this);
-    this.unflip = this.unflip.bind(this);
-  }
-
-  flip(e) {
-    e.preventDefault();
-    this.setState({ isFlipped: true});
-  }
-
-  unflip(e) {
-    e.preventDefault();
-    this.setState({ isFlipped: false});
-  }
- 
-  render() {
-    return (
-      <EntryContainer onMouseLeave={this.unflip} onMouseEnter={this.flip}>
-      <ReactCardFlip isFlipped={this.state.isFlipped} flipDirection="horizontal" style={{minHeight: '8em'}}>
-      <div key={'front'} style={{minHeight: '10em'}}>
-          <EntryThumb entry={this.props.picture} />
-          <Button color='secondary' style={{width: '100%'}} className="fixed-bottom" onClick={this.flip}>Options...</Button>
-        </div>
- 
-        <div key={'back'} style={{minHeight: '10em'}}>
-        <Button color='primary' style={{width: '100%'}} className="mb-4" >View</Button>
-        <Button color='primary' style={{width: '100%'}} >Update</Button>
-        <Button color='danger' style={{width: '100%' }} className="fixed-bottom">Delete</Button>
-        </div>
-      </ReactCardFlip>
-      </EntryContainer>
-    )
-  }
-}
 const SubmittedEntries = ({ show }) =>
   show.entries.map(entry => (
     <Col
@@ -173,8 +94,11 @@ const SubmittedEntries = ({ show }) =>
     title={entry.title}
     key={entry.id}
   >
-    <FlipCard picture={entry} style={{width: '100%', height: '100%'}}>
-    </FlipCard>
+    {/* <FlipCard picture={entry} style={{width: '100%', height: '100%'}}>
+    </FlipCard> */}
+
+    <DarkCard picture={entry} style={{width: '100%', height: '100%'}}>
+    </DarkCard>
         {/* <EntryContainer>
     <Flippy id="flippy"
     flipOnHover={false} // default false
